[math] Support Jax sparse COO (non-batched)
* Remove Backend.coordinates (unused and definition ambiguous)
