+++ tests/commit/math/test__ops.py
@@ -644,21 +644,6 @@ class TestMathFunctions(TestCase):
         self.assertIs(math.numpy(nat), nat)
         assert_close(math.numpy(math.tensor(nat)), nat)
 
-    def test_sparse(self):
-        idx = [[0, 1, 1],
-             [2, 0, 2]]
-        v = [3, 4, 5]
-        shape = (2, 3)
-        for backend in BACKENDS:
-            if backend.supports(Backend.sparse_coo_tensor):
-                with backend:
-                    idx_ = backend.transpose(backend.as_tensor(idx), [1, 0])
-                    matrix = backend.sparse_coo_tensor(idx_, v, shape)
-                    i_, v_ = backend.coordinates(matrix)
-                    self.assertIsInstance(i_, tuple, msg=backend.name)
-                    assert len(i_) == 2
-                    assert len(v) == 3
-
     def test_rename_dims(self):
         t = math.zeros(spatial(x=5, y=4))
         self.assertEqual(math.rename_dims(t, 'x', 'z').shape.get_type('z'), 'spatial')
